migrations/versions/f82fa88ddefc_.py
"""empty message

Revision ID: f82fa88ddefc
Revises: 726fd2d270e7
Create Date: 2021-07-15 18:11:36.234618

"""
from alembic import op
import sqlalchemy as sa


# revision identifiers, used by Alembic.
revision = 'f82fa88ddefc'
down_revision = '726fd2d270e7'
branch_labels = None
depends_on = None


def upgrade():
    op.create_table('Show',
    sa.Column('id', sa.Integer(), nullable=False),
    sa.Column('artist_id', sa.Integer(), nullable=False),
    sa.Column('venue_id', sa.Integer(), nullable=False),
    sa.Column('start_time', sa.DateTime(), nullable=False),
    sa.ForeignKeyConstraint(['artist_id'], ['Artist.id'], ),
    sa.ForeignKeyConstraint(['venue_id'], ['Venue.id'], ),
    sa.PrimaryKeyConstraint('id')
    )


def downgrade():
    op.drop_table('Show')
